Initial pass for parallel exception hierarchy. #4781
This work is the start of producing a Java-domain Throwable
hierarchy that matches the Ruby hierarchy, so that Ruby exceptions
can be caught using their actual Throwable type rather than using
RaiseException and checking the exceptions type manually.

A few changes of note:

* RubyException has most logic moved up to new
  AbstractRubyException.
* All RubyException now hold a reference to their RaiseException,
  created lazily once as needed.
  * Throw a Ruby exception using the form
    `throw ex.getRaiseException()`.
  * Provide construction logic specific to a given exception by
    overriding createRaiseException.
* The nativeException parameter is eliminated from all exception
  construction and initialization paths. It was unused.
