+++ core/src/main/java/org/jruby/runtime/backtrace/TraceType.java
@@ -3,7 +3,6 @@ package org.jruby.runtime.backtrace;
 import java.io.ByteArrayOutputStream;
 import java.io.FileDescriptor;
 import java.io.PrintStream;
-import java.util.Arrays;
 
 import org.jruby.Ruby;
 import org.jruby.RubyArray;
@@ -40,11 +39,10 @@ public class TraceType {
      * Get a normal Ruby backtrace, using the current Gather type.
      *
      * @param context
-     * @param nativeException
      * @return
      */
-    public BacktraceData getBacktrace(ThreadContext context, boolean nativeException) {
-        return gather.getBacktraceData(context, nativeException);
+    public BacktraceData getBacktrace(ThreadContext context) {
+        return gather.getBacktraceData(context);
     }
 
     /**
@@ -167,7 +165,7 @@ public class TraceType {
          * Full raw backtraces with all Java frames included.
          */
         RAW {
-            public BacktraceData getBacktraceData(ThreadContext context, StackTraceElement[] javaTrace, boolean nativeException) {
+            public BacktraceData getBacktraceData(ThreadContext context, StackTraceElement[] javaTrace) {
                 return new BacktraceData(
                         javaTrace,
                         BacktraceElement.EMPTY_ARRAY,
@@ -181,7 +179,7 @@ public class TraceType {
          * A backtrace with interpreted frames intact, but don't remove Java frames.
          */
         FULL {
-            public BacktraceData getBacktraceData(ThreadContext context, StackTraceElement[] javaTrace, boolean nativeException) {
+            public BacktraceData getBacktraceData(ThreadContext context, StackTraceElement[] javaTrace) {
                 return new BacktraceData(
                         javaTrace,
                         context.getBacktrace(),
@@ -195,7 +193,7 @@ public class TraceType {
          * A normal Ruby-style backtrace, but which includes any non-org.jruby frames
          */
         INTEGRATED {
-            public BacktraceData getBacktraceData(ThreadContext context, StackTraceElement[] javaTrace, boolean nativeException) {
+            public BacktraceData getBacktraceData(ThreadContext context, StackTraceElement[] javaTrace) {
                 return new BacktraceData(
                         javaTrace,
                         context.getBacktrace(),
@@ -209,7 +207,7 @@ public class TraceType {
          * Normal Ruby-style backtrace, showing only Ruby and core class methods.
          */
         NORMAL {
-            public BacktraceData getBacktraceData(ThreadContext context, StackTraceElement[] javaTrace, boolean nativeException) {
+            public BacktraceData getBacktraceData(ThreadContext context, StackTraceElement[] javaTrace) {
                 return new BacktraceData(
                         javaTrace,
                         context.getBacktrace(),
@@ -223,7 +221,7 @@ public class TraceType {
          * Normal Ruby-style backtrace, showing only Ruby and core class methods.
          */
         CALLER {
-            public BacktraceData getBacktraceData(ThreadContext context, StackTraceElement[] javaTrace, boolean nativeException) {
+            public BacktraceData getBacktraceData(ThreadContext context, StackTraceElement[] javaTrace) {
                 return new BacktraceData(
                         javaTrace,
                         context.getBacktrace(),
@@ -237,11 +235,10 @@ public class TraceType {
          * Gather backtrace data for a normal Ruby trace.
          *
          * @param context
-         * @param nativeException
          * @return
          */
-        public BacktraceData getBacktraceData(ThreadContext context, boolean nativeException) {
-            BacktraceData data = getBacktraceData(context, Thread.currentThread().getStackTrace(), nativeException);
+        public BacktraceData getBacktraceData(ThreadContext context) {
+            BacktraceData data = getBacktraceData(context, Thread.currentThread().getStackTrace());
 
             context.runtime.incrementBacktraceCount();
             if (RubyInstanceConfig.LOG_BACKTRACES) logBacktrace(data.getBacktrace(context.runtime));
@@ -264,7 +261,7 @@ public class TraceType {
                 useGather = INTEGRATED;
             }
 
-            BacktraceData data = useGather.getBacktraceData(context, javaTrace, false);
+            BacktraceData data = useGather.getBacktraceData(context, javaTrace);
 
             context.runtime.incrementBacktraceCount();
             if (RubyInstanceConfig.LOG_BACKTRACES) logBacktrace(data.getBacktrace(context.runtime));
@@ -272,7 +269,7 @@ public class TraceType {
             return data;
         }
 
-        public abstract BacktraceData getBacktraceData(ThreadContext context, StackTraceElement[] javaTrace, boolean nativeException);
+        public abstract BacktraceData getBacktraceData(ThreadContext context, StackTraceElement[] javaTrace);
     }
 
     public enum Format {
